refactor(toggle): updated component structure
