+++ packages/react/src/components/toggle/toggle-group.stories.tsx
@@ -1,0 +1,162 @@
+import type { Meta, StoryFn } from "@storybook/react"
+import { useState } from "react"
+import { Toggle, ToggleGroup } from "."
+import { PropsTable } from "../../../storybook/components"
+import { COLOR_SCHEMES } from "../../utils"
+import { For } from "../for"
+import { BoldIcon, ItalicIcon, UnderlineIcon } from "../icon"
+
+type Story = StoryFn<typeof Toggle>
+
+const meta: Meta<typeof Toggle> = {
+  component: Toggle,
+  title: "Components / ToggleGroup",
+}
+
+export default meta
+
+export const Basic: Story = () => {
+  return (
+    <ToggleGroup>
+      <Toggle aria-label="Toggle bold" icon={<BoldIcon />} value="bold" />
+      <Toggle aria-label="Toggle italic" icon={<ItalicIcon />} value="italic" />
+      <Toggle
+        aria-label="Toggle underline"
+        icon={<UnderlineIcon />}
+        value="underline"
+      />
+    </ToggleGroup>
+  )
+}
+
+export const Size: Story = () => {
+  return (
+    <PropsTable columns={["xs", "sm", "md", "lg"]} rows={COLOR_SCHEMES}>
+      {(column, row, key) => {
+        return (
+          <ToggleGroup key={key} colorScheme={row} size={column}>
+            <Toggle aria-label="Toggle bold" icon={<BoldIcon />} value="bold" />
+            <Toggle
+              aria-label="Toggle italic"
+              icon={<ItalicIcon />}
+              value="italic"
+            />
+            <Toggle
+              aria-label="Toggle underline"
+              icon={<UnderlineIcon />}
+              value="underline"
+            />
+          </ToggleGroup>
+        )
+      }}
+    </PropsTable>
+  )
+}
+
+export const Variant: Story = () => {
+  return (
+    <PropsTable
+      columns={["solid", "subtle", "surface", "outline"]}
+      rows={COLOR_SCHEMES}
+    >
+      {(column, row, key) => {
+        return (
+          <ToggleGroup key={key} colorScheme={row} variant={column}>
+            <Toggle aria-label="Toggle bold" icon={<BoldIcon />} value="bold" />
+            <Toggle
+              aria-label="Toggle italic"
+              icon={<ItalicIcon />}
+              value="italic"
+            />
+            <Toggle
+              aria-label="Toggle underline"
+              icon={<UnderlineIcon />}
+              value="underline"
+            />
+          </ToggleGroup>
+        )
+      }}
+    </PropsTable>
+  )
+}
+
+export const Disabled: Story = () => {
+  return (
+    <For each={["solid", "subtle", "surface", "outline"]}>
+      {(variant, index) => (
+        <ToggleGroup key={index} variant={variant} disabled>
+          <Toggle aria-label="Toggle bold" icon={<BoldIcon />} value="bold" />
+          <Toggle
+            aria-label="Toggle italic"
+            icon={<ItalicIcon />}
+            value="italic"
+          />
+          <Toggle
+            aria-label="Toggle underline"
+            icon={<UnderlineIcon />}
+            value="underline"
+          />
+        </ToggleGroup>
+      )}
+    </For>
+  )
+}
+
+export const Readonly: Story = () => {
+  return (
+    <For each={["solid", "subtle", "surface", "outline"]}>
+      {(variant, index) => (
+        <ToggleGroup key={index} variant={variant} readOnly>
+          <Toggle aria-label="Toggle bold" icon={<BoldIcon />} value="bold" />
+          <Toggle
+            aria-label="Toggle italic"
+            icon={<ItalicIcon />}
+            value="italic"
+          />
+          <Toggle
+            aria-label="Toggle underline"
+            icon={<UnderlineIcon />}
+            value="underline"
+          />
+        </ToggleGroup>
+      )}
+    </For>
+  )
+}
+
+export const CustomControl: Story = () => {
+  const [singleValue, setSingleValue] = useState<string | undefined>(undefined)
+  const [multiValue, setMultiValue] = useState<string[]>([])
+
+  return (
+    <>
+      <ToggleGroup value={singleValue} onChange={setSingleValue}>
+        <Toggle aria-label="Toggle bold" icon={<BoldIcon />} value="bold" />
+        <Toggle
+          aria-label="Toggle italic"
+          icon={<ItalicIcon />}
+          value="italic"
+        />
+        <Toggle
+          aria-label="Toggle underline"
+          icon={<UnderlineIcon />}
+          value="underline"
+        />
+      </ToggleGroup>
+
+      <ToggleGroup value={multiValue} onChange={setMultiValue}>
+        <Toggle aria-label="Toggle bold" icon={<BoldIcon />} value="bold" />
+        <Toggle
+          aria-label="Toggle italic"
+          icon={<ItalicIcon />}
+          value="italic"
+        />
+        <Toggle
+          aria-label="Toggle underline"
+          icon={<UnderlineIcon />}
+          value="underline"
+        />
+      </ToggleGroup>
+    </>
+  )
+}
