+++ packages/react/src/index.ts
@@ -107,7 +107,7 @@ export * from "./components/status"
 export * from "./components/tag"
 export * from "./components/text"
 export * from "./components/textarea"
-// export * from "./components/toggle"
+export * from "./components/toggle"
 // export * from "./components/tooltip"
 export * from "./components/visually-hidden"
 export * from "./components/wrap"
