--- packages/react/src/legacy-components/toggle/index.ts
@@ -1,4 +1,0 @@
-export { Toggle } from "./toggle"
-export type { ToggleProps } from "./toggle"
-export { ToggleGroup } from "./toggle-group"
-export type { ToggleGroupProps } from "./toggle-group"
